Puzzle Game
It is a game that tries to make the original picture by shifting the randomly distributed picture pieces to the right, left, down and up. Each movement is recorded and activated by the timer, which takes 1 minute. With the buttons, the tracks were shuffled again or the game was stopped or used to exit the game.